Revert "Merge branch 'master' of https://github.com/DarkMiaou/Tales-of-Hivnir"
This reverts commit 3be62d8c02bc0575859b628dfacbad78304415de, reversing
changes made to e201aeb50096af27a64171be7b5bef41cb9c55a6.

+++ Tiles-Of-Hivnir/Assets/Scripts/Combat/Enemy/EnemyCombat.cs
@@ -36,7 +36,7 @@ public class EnemyCombat : MonoBehaviour
         }*/
         if (isinside == true)
         {  
-            SaveData.instance.monstercurrenthealth -= (int)SaveData.instance.damage;
+            SaveData.instance.monstercurrenthealth -= SaveData.instance.damage;
             Debug.Log("Attaque réalisé");
         }
